feat: Extract channel management tools to dedicated module (Phase 5)
- Move all 9 channel management tools to tools/channel_management.py:
  - create_channel
  - archive_channel
  - set_channel_topic
  - set_channel_purpose
  - join_channel
  - leave_channel
  - invite_to_channel
  - remove_from_channel
  - list_channel_members

- Implement register_tools() pattern for module registration
- Update imports and registration in main server file
- Update tools/__init__.py to include channel_management module
- Reduce main server file by ~438 lines

This completes Phase 5 of the modularization refactor, maintaining
backward compatibility while improving code organization.

Fixes #21
