+++ src/slack_mcp/tools/__init__.py
@@ -3,5 +3,6 @@
 # Import all tool modules to ensure registration
 from . import core
 from . import message_management
+from . import channel_management
 
-__all__ = ["core", "message_management"]
+__all__ = ["core", "message_management", "channel_management"]
